api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sse-starlette==2.1.3
httpx==0.27.2
firebase-admin==6.5.0
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.71.1
google-auth==2.34.0
anthropic[vertex]==0.39.0
